% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/08_fourier.R
\name{basis_fourier}
\alias{basis_fourier}
\title{Creating Fourier Basis Functions}
\usage{
basis_fourier(range = c(0, 1), nbasis = 3, period = diff(range),
  dropind = NULL)
}
\arguments{
\item{range}{A length-two numeric vector to define the interval on which
basis functions can be evaluated. Default is \code{c(0, 1)}.}

\item{nbasis}{The total number of basis functions, including the ones that need
to be dropped. \code{nbasis} should be an odd number. If an
even number is specified, it will be increased by 1 automatically.
Default is 3.}

\item{period}{The length of the cycle that Fourier basis functions repeat
themselves. Default is the width of \code{range}.}

\item{dropind}{Indices of basis functions that need to be dropped. Default is
               \code{NULL}, meaning no basis will be dropped.}
}
\value{
A \code{\link[=fourier+-class]{fourier+}} object representing the
        basis functions.
}
\description{
This function constructs a \code{\link[=fourier+-class]{fourier+}} object
that represents a series of Fourier basis functions.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

