% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/20_bifd.R
\docType{class}
\name{bifd+-class}
\alias{bifd+-class}
\title{Bivariate Function Using Basis Expansion}
\description{
This class defines bivariate functions that can be expressed by two sets
of basis functions and the associated coefficient matrix. It takes the following
form:
}
\details{
\deqn{K(s,t)=\sum_{i,j} a_{ij}f_i(s)g_j(t)}{K(s, t) = sum_{i,j} a_{ij} * f_i(s) * g_j(t)}

Here \eqn{K(s, t)} is the bivariate function, \eqn{f_i} and \eqn{g_j} are
two basis systems, and \eqn{A = (a_{ij})} is the coefficient matrix.
}
\section{Slots}{

\describe{
\item{\code{sbasis,tbasis}}{Basis objects of class \code{\link[=basis+-class]{basis+}},
not necessarily of the same type (for example, one can be
\code{\link[=bspline+-class]{bspline+}} and the other be
\code{\link[=fourier+-class]{fourier+}}).}

\item{\code{coefs}}{A matrix of dimension \code{m} by \code{n} where \code{m}
            is the number of functions in \code{sbasis}, and \code{n}
            is the number of functions in \code{tbasis}.}
}}

