% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/05_basis.R
\docType{class}
\name{bspline+-class}
\alias{bspline+-class}
\title{B-spline Basis Functions}
\description{
The "\code{bspline+}" class represents B-spline basis functions.
}
\section{Slots}{

\describe{
\item{\code{range}}{A length-two numeric vector to define the interval on which
basis functions can be evaluated. Default is \code{c(0, 1)}.}

\item{\code{nbasis}}{The total number of basis functions, including the ones that need
to be dropped. Default is 4.}

\item{\code{dropind}}{Indices of basis functions that need to be dropped. Default is
\code{numeric(0)} (no basis will be dropped).}

\item{\code{ncoef}}{The actual number of basis functions to represent functional data,
satisfying the condition \code{ncoef == nbasis - length(dropind)}.}

\item{\code{degree}}{Degree of the B-splines. Default is 3, standing for cubic splines.}

\item{\code{knots}}{A numeric vector giving the inner knots. Must satisfy
            \code{nbasis == degree + length(knots) + 1}.}
}}
\section{Extends}{

Class \code{\link[=basis+-class]{basis+}}
}

