% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/15_fd.R
\name{c.fd+}
\alias{c.fd+}
\title{Combining Multiple Univariate Functional Data Objects}
\usage{
\method{c}{`fd+`}(x, ...)
}
\arguments{
\item{x,\dots}{\code{\link[=fd+-class]{fd+}} objects to be combined. They should
               have the same basis functions.}
}
\value{
An \code{\link[=fd+-class]{fd+}} object that contains all the functions
        in the list.
}
\description{
This function combines a series of \code{\link[=fd+-class]{fd+}} objects
that share the same basis into a single one.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

