% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/15_fd.R
\docType{class}
\name{fd+-class}
\alias{fd+-class}
\title{Univariate Functional Data Object}
\description{
A univariate functional data object, of the class \code{fd+}, can be
thought of as a vector whose elements are univariate functions
that can be expressed as linear combination of a system of basis functions.
Functions in one object are assumed to share the same basis, so an
\code{fd+} object is composed of one \code{\link[=basis+-class]{basis+}}
object and the corresponding coefficients.
}
\section{Slots}{

\describe{
\item{\code{basis}}{Basis object of class \code{\link[=basis+-class]{basis+}}.}

\item{\code{coefs}}{A matrix of dimension \code{n} by \code{p} where \code{n}
            is the number of functions in this object, and \code{p}
            is the number of basis functions in \code{basis}.
            The \code{i}-th row of \code{coefs} contains the coefficients
            for the \code{i}-th function in the object.}
}}

