% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/15_fd.R
\name{fd_new}
\alias{fd_new}
\title{Creating A Functional Data Object}
\usage{
fd_new(coefs, basis)
}
\arguments{
\item{coefs}{A matrix whose each row gives the coefficients of basis functions
for each curve in this object. If \code{coefs} is a vector, it
is treated as a matrix of one row.}

\item{basis}{Basis of this functional data object. Should be of class
             \code{\link[=basis+-class]{basis+}} or one of its subclasses
             (\code{\link[=bspline+-class]{bspline+}} and
             \code{\link[=fourier+-class]{fourier+}}).}
}
\value{
An \code{\link[=fd+-class]{fd+}} object with the given basis and coefficients.
}
\description{
This function constructs an \code{\link[=fd+-class]{fd+}} object
that represents a set of functional data.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

