% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/08_fourier.R, R/10_bspline.R, R/15_fd.R, R/20_bifd.R
\docType{methods}
\name{feval-methods}
\alias{feval}
\alias{feval,bifd+,numeric-method}
\alias{feval,bspline+,numeric-method}
\alias{feval,fd+,numeric-method}
\alias{feval,fourier+,numeric-method}
\alias{feval-methods}
\title{Evaluating A Functional Data Object}
\usage{
feval(f, x, ...)

## For bifd+
# feval(f, x, y, ...)

\S4method{feval}{`fourier+`,numeric}(f, x, ...)

\S4method{feval}{`bspline+`,numeric}(f, x, ...)

\S4method{feval}{`fd+`,numeric}(f, x, ...)

\S4method{feval}{`bifd+`,numeric}(f, x, y, ...)
}
\arguments{
\item{f}{An object of class \code{\link[=basis+-class]{basis+}},
\code{\link[=fd+-class]{fd+}}, or \code{\link[=bifd+-class]{bifd+}}.}

\item{x}{Values that \code{f} is evaluated on.}

\item{y}{Only used when \code{f} is of class \code{\link[=bifd+-class]{bifd+}},
in which case \code{f} is evaluated on two arguments, \code{x} and
\code{y}.}

\item{\dots}{Additional parameters, currently unused.}
}
\value{
See \strong{Method (...)} sections for the actual implementations of
        this generic function.
}
\description{
This generic function provides methods to evaluate different types of
functional data (basis function, univariate function, bivariate function)
at specified argument values.
}
\section{Method (fourier+, numeric)}{

\tabular{lcl}{
  \code{f}  \tab - \tab  A \code{\link[=fourier+-class]{fourier+}} object. \cr
  \code{x}  \tab - \tab  A numeric vector.
}

\code{feval(f, x)} returns a matrix \code{R} of
\code{f@ncoef} rows and \code{length(x)} columns, with \code{R[i, j]}
representing the value of the \code{i}-th basis function evaluated on \code{x[j]}.
}

\section{Method (bspline+, numeric)}{

\tabular{lcl}{
  \code{f}  \tab - \tab  A \code{\link[=bspline+-class]{bspline+}} object. \cr
  \code{x}  \tab - \tab  A numeric vector.
}

\code{feval(f, x)} returns a matrix \code{R} of
\code{f@ncoef} rows and \code{length(x)} columns, with \code{R[i, j]}
representing the value of the \code{i}-th basis function evaluated on \code{x[j]}.
}

\section{Method (fd+, numeric)}{

\tabular{lcl}{
  \code{f}  \tab - \tab  An \code{\link[=fd+-class]{fd+}} object. \cr
  \code{x}  \tab - \tab  A numeric vector.
}

\code{feval(f, x)} returns a matrix \code{R} of
\code{n} rows and \code{length(x)} columns where \code{n} is the number of
function curves in \code{f}. \code{R[i, j]}
equals the value of the \code{i}-th function evaluated on \code{x[j]}.
}

\section{Method (bifd+, numeric, numeric)}{

\tabular{lcl}{
  \code{f}     \tab - \tab  A \code{\link[=bifd+-class]{bifd+}} object. \cr
  \code{x, y}  \tab - \tab  Numeric vectors.
}

\code{feval(f, x, y)} returns a matrix \code{R} of
\code{length(x)} rows and \code{length(y)} columns, with \code{R[i, j]}
equal to the value of \code{f(x[i], x[j])}.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

