% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/08_fourier.R, R/10_bspline.R, R/15_fd.R, R/20_bifd.R
\docType{methods}
\name{feval}
\alias{feval}
\alias{feval,bifd+,numeric-method}
\alias{feval,bspline+,numeric-method}
\alias{feval,fd+,numeric-method}
\alias{feval,fourier+,numeric-method}
\title{Evaluating Various Types of Functional Data Objects}
\usage{
feval(f, x, ...)

## For bifd+
# feval(f, x, y, ...)
}
\arguments{
\item{f}{An object of class \code{\link[=basis+-class]{basis+}},
\code{\link[=fd+-class]{fd+}}, or \code{\link[=bifd+-class]{bifd+}}.}

\item{x}{Values that \code{f} is evaluated on.}

\item{y}{Only used when \code{f} is of class \code{\link[=bifd+-class]{bifd+}}.}

\item{\dots}{Additional parameters, currently unused.}
}
\value{
See section \strong{Methods} for the actual implementations of
        this generic function.
}
\description{
Evaluating Various Types of Functional Data Objects
}
\section{Methods (by class)}{
\itemize{
\item \code{f = fourier+,x = numeric}: Evaluating a Fourier Basis Function Object

\code{x} is a numeric vector. \code{feval(f, x)} returns a matrix \code{R} of
\code{f@ncoef} rows and \code{length(x)} columns, with \code{R[i, j]}
representing the value of the \code{i}-th basis function evaluated on \code{x[j]}.

\item \code{f = bspline+,x = numeric}: Evaluating a B-spline Basis Function Object

\code{x} is a numeric vector. \code{feval(f, x)} returns a matrix \code{R} of
\code{f@ncoef} rows and \code{length(x)} columns, with \code{R[i, j]}
representing the value of the \code{i}-th basis function evaluated on \code{x[j]}.

\item \code{f = fd+,x = numeric}: Evaluating a Functional Data Object

\code{x} is a numeric vector. \code{feval(f, x)} returns a matrix \code{R} of
\code{n} rows and \code{length(x)} columns where \code{n} is the number of
function curves in \code{f}. \code{R[i, j]}
equals the value of the \code{i}-th function evaluated on \code{x[j]}.

\item \code{f = bifd+,x = numeric}: Evaluating a Bivariate Functional Data Object

\code{x} and \code{y} are numeric vectors.
\code{feval(f, x, y)} returns a matrix \code{R} of
\code{length(x)} rows and \code{length(y)} columns, with \code{R[i, j]}
equal to the value of \code{f(x[i], [j])}.
}}

