% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/15_fd.R
\docType{methods}
\name{mean-methods}
\alias{mean,fd+-method}
\alias{mean-methods}
\title{Mean Function of Univariate Functinoal Data}
\usage{
\S4method{mean}{`fd+`}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=fd+-class]{fd+}} object.}

\item{\dots}{Currently unused.}
}
\value{
An \code{\link[=fd+-class]{fd+}} object that contains the mean function.
}
\description{
This function calculates the mean curve of a set of univariate functional data.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

