% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/08_fourier.R, R/10_bspline.R
\docType{methods}
\name{penmat-methods}
\alias{penmat}
\alias{penmat,bspline+,numeric-method}
\alias{penmat,fourier+,numeric-method}
\alias{penmat-methods}
\title{Calculating the Basis Penalty Matrix}
\usage{
penmat(basis, penalty, ...)

\S4method{penmat}{`fourier+`,numeric}(basis, penalty, ...)

\S4method{penmat}{`bspline+`,numeric}(basis, penalty, ...)
}
\arguments{
\item{basis}{An object inherited from \code{\link[=basis+-class]{basis+}}.
Can be of class \code{\link[=bspline+-class]{bspline+}} or
\code{\link[=fourier+-class]{fourier+}}.}

\item{penalty}{An integer giving the order of derivative.}

\item{\dots}{Currently unused.}
}
\value{
A symmetric non-negative definite penalty matrix.
}
\description{
This function calculates the penalty matrix \eqn{R} of a basis object
(for example of class \code{\link[=bspline+-class]{bspline+}} or
\code{\link[=fourier+-class]{fourier+}}). The element
\eqn{R_{ij}}{R[i, j]} is equal
to the inner product of \eqn{f_i^{(k)}}{f_i^(k)} and
\eqn{f_j^{(k)}}{f_j^(k)}, where \eqn{f_i^{(k)}}{f_i^(k)} is the \eqn{k}-th
derivative of the \eqn{i}-th basis.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

