% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/05_basis.R, R/08_fourier.R, R/10_bspline.R, R/15_fd.R, R/20_bifd.R
\docType{methods}
\name{plot-methods}
\alias{plot,basis+,missing-method}
\alias{plot,bifd+,missing-method}
\alias{plot,bspline+,missing-method}
\alias{plot,fd+,missing-method}
\alias{plot,fourier+,missing-method}
\alias{plot-methods}
\alias{plot3d,bifd+-method}
\title{Visualizing A Functional Data Object}
\usage{
\S4method{plot}{`basis+`,missing}(x, y, ...)

\S4method{plot}{`fourier+`,missing}(x, y, ...)

\S4method{plot}{`bspline+`,missing}(x, y, ...)

\S4method{plot}{`fd+`,missing}(x, y, ...)

\S4method{plot}{`bifd+`,missing}(x, y, ...)

\S4method{plot3d}{`bifd+`}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link[=basis+-class]{basis+}},
\code{\link[=fd+-class]{fd+}}, or \code{\link[=bifd+-class]{bifd+}}.}

\item{y}{Unused and should be missing. For compatibility with the signature
of the generic function.}

\item{\dots}{Additional graphical parameters passed to \code{matplot()},
       \code{persp()} or \code{rgl::persp3d()}.}
}
\description{
This generic function provides methods to visualize different types of
functional data (basis function, univariate function, bivariate function).
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

