% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/05_basis.R, R/15_fd.R
\docType{methods}
\name{[-methods}
\alias{[,basis+,numeric,missing-method}
\alias{[,fd+,numeric,missing-method}
\alias{[-methods}
\title{Subsetting A Functional Data Object}
\usage{
\S4method{[}{`basis+`,numeric,missing}(x, i)

\S4method{[}{`fd+`,numeric,missing}(x, i)
}
\arguments{
\item{x}{An object of class \code{\link[=basis+-class]{basis+}}
or \code{\link[=fd+-class]{fd+}}.}

\item{i}{An integer vector giving the indices of functions that are going
to be extracted.}

\item{\dots}{Additional parameters, currently unused.}
}
\value{
An object of the same class as \code{x} containing the extracted functions.
}
\description{
This generic function provides methods to subset a functional data
object (basis function or univariate function).
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

