% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/05_basis.R, R/15_fd.R, R/20_bifd.R
\name{wrap}
\alias{wrap}
\alias{wrap.basisfd}
\alias{wrap.bifd}
\alias{wrap.fd}
\title{Converting Objects in the 'fda' Package into their Counterparts in 'fdaplus'}
\usage{
wrap(obj, ...)

\method{wrap}{basisfd}(obj, ...)

\method{wrap}{fd}(obj, ...)

\method{wrap}{bifd}(obj, ...)
}
\arguments{
\item{obj}{Object to be converted, whose type should be "\code{basisfd}",
"\code{fd}" or "\code{bifd}".}

\item{\dots}{Currently unused.}
}
\value{
Below gives the conversion result:
\itemize{
  \item \code{basisfd} objects will be converted into
        \code{\link[=basis+-class]{basis+}} objects
  \item \code{fd} objects will be converted into
        \code{\link[=fd+-class]{fd+}} objects
  \item \code{bifd} objects will be converted into
        \code{\link[=bifd+-class]{bifd+}} objects
}
}
\description{
This function can convert objects in the \pkg{fda} package
into S4 objects that have been redesigned in the \pkg{fdaplus} package.
}
\section{Methods (by class)}{
\itemize{
\item \code{basisfd}: Converting "basisfd" objects

\item \code{fd}: Converting "fd" objects

\item \code{bifd}: Converting "bifd" objects
}}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

