% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_generic.R, R/15_fd.R, R/20_bifd.R
\docType{methods}
\name{arithmetic-methods}
\alias{*,bifd+,numeric-method}
\alias{*,fd+,numeric-method}
\alias{*,numeric,bifd+-method}
\alias{*,numeric,fd+-method}
\alias{+,bifd+,bifd+-method}
\alias{+,fd+,fd+-method}
\alias{-,bifd+,bifd+-method}
\alias{-,fd+,fd+-method}
\alias{/,bifd+,numeric-method}
\alias{/,fd+,numeric-method}
\alias{arithmetic-methods}
\title{Arithmetic Operations of Functional Data Objects}
\usage{
\S4method{*}{`fd+`,numeric}(e1, e2)

\S4method{*}{numeric,`fd+`}(e1, e2)

\S4method{/}{`fd+`,numeric}(e1, e2)

\S4method{+}{`fd+`,`fd+`}(e1, e2)

\S4method{-}{`fd+`,`fd+`}(e1, e2)

\S4method{*}{`bifd+`,numeric}(e1, e2)

\S4method{*}{numeric,`bifd+`}(e1, e2)

\S4method{/}{`bifd+`,numeric}(e1, e2)

\S4method{+}{`bifd+`,`bifd+`}(e1, e2)

\S4method{-}{`bifd+`,`bifd+`}(e1, e2)
}
\arguments{
\item{e1,e2}{Numeric scalars, or functional data objects of class
             \code{\link[=fd+-class]{fd+}}
             or \code{\link[=bifd+-class]{bifd+}}.}
}
\description{
Simple arithmetic operations of \code{\link[=fd+-class]{fd+}}
and \code{\link[=bifd+-class]{bifd+}} objects.
}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

