% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/20_bifd.R
\name{bifd_new}
\alias{bifd_new}
\title{Creating A Bivariate Function Using Basis Expansion}
\usage{
bifd_new(coefs, sbasis, tbasis = sbasis)
}
\value{
An \code{\link[=bifd+-class]{bifd+}} object with the given
        bases and coefficients.
}
\description{
This function constructs a \code{\link[=bifd+-class]{bifd+}} object
that represents a bivariate function.
}
\section{Slots}{

\describe{
\item{\code{coefs}}{The coefficient matrix.}

\item{\code{sbasis,tbasis}}{Basis objects of class \code{\link[=basis+-class]{basis+}},
                    not necessarily of the same type (for example, one can be
                    \code{\link[=bspline+-class]{bspline+}} and the other be
                    \code{\link[=fourier+-class]{fourier+}}).}
}}
\author{
Yixuan Qiu <\url{http://statr.me/}>
}

